requirements.txt
aiohttp==3.8.3
azure-core==1.26.0
azure-data-tables==12.4.1
fastapi==0.85.1
pydantic==1.10.2
uvicorn==0.18.3
gunicorn==20.1.0